+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-07-15
+
+### Added
+
+- A new variable (`enable_prometheus_server`) has been added that can be used to
+  enable the prometheus server endpoint that can be used to gain some metrics.
+
+### Fixed
+
+- A "node anti affinity" term has been added to the Daemon set to ensure that
+  workloads will never be scheduled on Fargate nodes.
+
+### Updated
+
+- Default version of the aws-node-termination-handler has been set to v1.6.1.
+
 ## [1.1.0] - 2020-05-11
 
 ### Added